Final degree project of my bachelor's degree in Mathematics.
The special linear group of order 2, SL 2 (R), is the group of real matrices of order 2 whose determinant is 1 with matrix multiplication as group operation. We will carry out a study of this group from the approach of Lie groups. Some of the results that we K = R or C. The main goal of the work is the study of the Cartan subgroups of SL 2 (R). In order to define and find these will obtain will be generalizable to SL n (K) where subgroups, we will do a previous study in chapter 2 of topological groups (of which Lie groups are a special case) and Lie algebras. In chapter 3 we will focus in the group SL 2 (R) and its Lie algebra sl 2 (R) . We will look at the properties of point of view of group theory. As a main result, we have the Iwasawa decomposition of SL 2 (R) SL 2 (R) from the Iwasawa Decomposition , SL 2 (R) SL 2 (R) that into three subgroups obtained from the action of on the upper complex half-plane. We will also study the properties of will allow us to find its Cartan subgroups. Finally, we will see that up to conjugation, there are only two Cartan subgroups of subgroups covers almost all SL 2 (R) . SL 2 (R) . We will see that the union of these  We will study the similarities and differences of these two classes of Cartan subgroups. In particular, we will see that one class is syndetic in SL 2 (R) (i.e., finitely many translates cover the whole group) and the other class is not.